pyproject.toml
[project]
name = "sql-server-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=2.1.4",
    "fastapi>=0.119.0",
    "openai>=2.3.0",
    "pydantic>=2.12.2",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.37.0",
]
